Fixed oversight where anim id gets mangled and accepted

+++ Crayon/include/crayon/memory.h
@@ -66,7 +66,7 @@ uint8_t crayon_memory_clone_palette(crayon_palette_t *original, crayon_palette_t
 	// If T is set, we ignore the flip, frames and Left colour bits
 // If making an untextured array, set the spritesheet pointer to NULL
 uint8_t crayon_memory_init_sprite_array(crayon_sprite_array_t *sprite_array, crayon_spritesheet_t *ss,
-	uint8_t animation_id, crayon_palette_t *pal, uint16_t list_size, uint8_t frames_used, uint8_t options,
+	int16_t animation_id, crayon_palette_t *pal, uint16_t list_size, uint8_t frames_used, uint8_t options,
 	uint8_t filter, uint8_t set_defaults);
 
 // Note, we assume dest isn't initialised
